% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{virulent_predict}
\alias{virulent_predict}
\title{Predict virulence using a trained SVM model}
\usage{
virulent_predict(new_data, format = "Normalize")
}
\arguments{
\item{new_data}{Expression data (features as rows, samples as columns)}

\item{format}{Preprocessing method (default: "Normalize")}
}
\value{
Data frame with predictions
}
\description{
Predict virulence using a trained SVM model
}
